Added handling of TABLEs (especially for Freeze Frames)
Freeze Frames, especially 0x71, include regular MWBs as response parameters. They are implemented as a TABLE-KEY and TABLE-STRUCT pair, which is also the case for some other types of special Freeze Frames.

This is also how MWBs are stored, so the scripts for dumping and parsing them have been changed to parse them automatically instead of manually extracting fields from the objects.